agents/skincare-intel/requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
certifi>=2024.0.0
truststore>=0.9.0
